Moved logic of panel/panel.py to base level.
Now single file `panel_view.py`.
Also moved `panel.sublime-syntax` to 'resources/'.
Panel seems no longer updating properly after those changes.
Also the color scheme of the panel has changed. (Font: yellow -> white).
Bound to squash
