+++ resources/panel.sublime-syntax
